accordionItem {shinydashboardPlus}	R Documentation

AdminLTE2 accordion item

Description

Create an accordion item to put inside an accordion container

Usage

accordionItem(..., id, title = NULL, color = NULL, collapsed = TRUE)

Arguments

...
text to write in the item.

id
unique item id.

title
item title.

color
item color: see here for a list of valid colors https://adminlte.io/themes/AdminLTE/pages/UI/general.html.

collapsed
Whether to expand or collapse the item. TRUE by default. Set it to FALSE if you want to expand it.

Author(s)

David Granjon, dgranjon@ymail.com

[Package shinydashboardPlus version 0.7.0 Index]