AdminLTE2 dashboard right sidebar

This creates a right sidebar.

dashboardControlbar(
  ...,
  inputId = NULL,
  skin = "dark",
  disable = FALSE,
  collapsed = TRUE,
  overlay = TRUE,
  width = 230,
  .items = NULL
)

Arguments

...
slot for rightSidebarTabContent. Not compatible with .items.

inputId
To acces the current state of the controlbar. Open is TRUE, closed is FALSE. NULL by default.

skin
background color: "dark" or "light".

disable
If TRUE, the sidebar will be disabled.

collapsed
Whether the control bar on the right side is collapsed or not at start. TRUE by default.

overlay
Whether the sidebar covers the content when expanded. Default to TRUE.

width
Sidebar width in pixels. Numeric value expected. 230 by default.

.items
Pass element here if you do not want to embed them in panels. Not compatible with ...

Note

Until a maximum of 5 rightSidebarTabContent inside! AdminLTE 2 does not support more panels.

Author

David Granjon, dgranjon@ymail.com

Examples

if (interactive()) {
 library(shiny)
 library(shinydashboard)
 library(shinydashboardPlus)
 shinyApp(
   ui = dashboardPage(
     header = dashboardHeader(),
     sidebar = dashboardSidebar(),
     body = dashboardBody(),
     controlbar = dashboardControlbar(
      skin = "dark",
      controlbarMenu(
       id = "menu",
       controlbarItem(
        "Tab 1",
        "Welcome to tab 1"
       ),
       controlbarItem(
        "Tab 2",
        "Welcome to tab 2"
       )
      )
     ),
     title = "Right Sidebar"
   ),
   server = function(input, output) { }
 )
}