dashboardPagePlus {shinydashboardPlus}	R Documentation

Dashboard Page with a right sidebar

Description

This creates a dashboard page for use in a Shiny app.

Usage

dashboardPagePlus(
  header,
  sidebar,
  body,
  rightsidebar = NULL,
  footer = NULL,
  title = NULL,
  skin = c("blue", "blue-light", "black", "black-light", "purple", "purple-light",
    "green", "green-light", "red", "red-light", "yellow", "yellow-light", "midnight"),
  collapse_sidebar = FALSE,
  sidebar_background = NULL,
  sidebar_fullCollapse = FALSE,
  enable_preloader = FALSE,
  loading_duration = 2,
  md = FALSE
)

Arguments

header
A header created by dashboardHeaderPlus.

sidebar
A sidebar created by dashboardSidebar.

body
A body created by dashboardBody.

rightsidebar
A right sidebar created by rightSidebar. NULL by default.

footer
A footer created by dashboardFooter.

title
A title to display in the browser's title bar. If no value is provided, it will try to extract the title from the dashboardHeaderPlus.

skin
A color theme. See https://adminlte.io/themes/AdminLTE/pages/UI/general.html.

collapse_sidebar
Whether to collapse the left sidebar. TRUE by default.

sidebar_background
Main sidebar background color: either "light" or NULL. NULL by default.

sidebar_fullCollapse
Whether to fully collapse the sidebar as with shinydashboard. FALSE by default.

enable_preloader
Whether to enable a page loader. FALSE by default.

loading_duration
Loader duration in seconds. 2s by default.

md
Whether to enable material design. Experimental...

See Also

dashboardHeaderPlus, dashboardSidebar, dashboardBody.

Examples

if (interactive()) {
library(shiny)
library(shinydashboard)
library(shinydashboardPlus)

shinyApp(
  ui = dashboardPagePlus(
    header = dashboardHeaderPlus(
     enable_rightsidebar = TRUE,
     rightSidebarIcon = "gears"
    ),
    sidebar = dashboardSidebar(),
    body = dashboardBody(),
    rightsidebar = rightSidebar(),
    title = "DashboardPage"
  ),
  server = function(input, output) { }
)
}

[Package shinydashboardPlus version 0.7.5 Index]