dropdownBlock {shinydashboardPlus}	R Documentation

Create a dropdown block to place in a dashboard header

Description

Create a dropdown block to place in a dashboard header

Usage

dropdownBlock(..., id, icon = NULL, title = NULL, badgeStatus = "danger")

Arguments

...
Items to put in the menu.

id
Dropdown block id.

icon
An icon to display in the header. Expect icon.

title
Dropdown block title.

badgeStatus
Dropdown badge status.

See Also

dashboardHeader for example usage.

[Package shinydashboardPlus version 2.0.0 Index]