Create a dropdown menu to place in a dashboard header

Create a dropdown menu to place in a dashboard header

dropdownMenu(..., type = c("messages", "notifications", "tasks"),
  badgeStatus = "primary", icon = NULL, headerText = NULL, .list = NULL)

Arguments

...
Items to put in the menu. Typically, message menus should contain messageItems, notification menus should contain notificationItems, and task menus should contain taskItems.

type
The type of menu. Should be one of "messages", "notifications", "tasks".

badgeStatus
The status of the badge which displays the number of items in the menu. This determines the badge's color. Valid statuses are listed in validStatuses. A value of NULL means to not display a badge.

icon
An icon to display in the header. By default, the icon is automatically selected depending on type, but it can be overriden with this argument.

headerText
An optional text argument used for the header of the dropdown menu (this is only visible when the menu is expanded). If none is provided by the user, the default is "You have x messages," where x is the number of items in the menu (if the type is specified to be "notifications" or "tasks," the default text shows "You have x notifications" or "You have x tasks," respectively).

.list
An optional list containing items to put in the menu Same as the ... arguments, but in list format. This can be useful when working with programmatically generated items.

See also