loadingState {shinydashboardPlus}	R Documentation

AdminLTE2 loading state element

Description

When a section is still work in progress or a computation is running

Usage

loadingState()

Note

Loading state can be programmatically used when a conputation is running for instance.

Author(s)

David Granjon, dgranjon@ymail.com

Examples

if (interactive()) {
 library(shiny)
 library(shinydashboard)
 library(shinydashboardPlus)
 
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     box(
      title = "loading spinner",
      loadingState()
      )
    ),
    title = "Loading State"
  ),
  server = function(input, output) { }
 )
}

[Package shinydashboardPlus version 2.0.0 Index]