menuIcon {shinydashboardPlus}	R Documentation

AdminLTE2 menu icon item

Description

Create a nice menu icon to insert in a rightsidebarMenuItem.

Usage

menuIcon(name, color)

Arguments

name
Icon name.

color
background color: see here for a list of valid colors https://adminlte.io/themes/AdminLTE/pages/UI/general.html.

Author(s)

David Granjon, dgranjon@ymail.com

[Package shinydashboardPlus version 0.7.0 Index]