AdminLTE2 timeline item

Create a timeline item

timelineItem(
  ...,
  icon = NULL,
  color = NULL,
  time = NULL,
  title = NULL,
  border = TRUE,
  footer = NULL
)

Arguments

...
any element such as timeLineItemMedia ...

icon
item icon. Expect icon.

color
item color: see here for a list of valid colors https://adminlte.io/themes/AdminLTE/pages/UI/general.html.

time
item date or time.

title
item title.

border
Whether to display a border between the header and the body. TRUE by default.

footer
item footer if any.

Author

David Granjon, dgranjon@ymail.com