userOutput {shinydashboardPlus}	R Documentation

Create a dynamic user output (client side)

Description

This can be used as a placeholder for dynamically-generated dashboardUser.

Usage

userOutput(id, tag = shiny::tags$li)

Arguments

id
Output variable name.

tag
A tag function, like tags$li or tags$ul.

See Also

renderUser for the corresponding server side function and examples.

Other user outputs: renderUser()

[Package shinydashboardPlus version 2.0.0 Index]