AdminLTE2 widget user box

Create widget user box

widgetUserBox(..., title = NULL, subtitle = NULL, type = NULL,
  background = FALSE, backgroundUrl = NULL, src = NULL, color = NULL,
  footer = NULL, footer_padding = TRUE, width = 6, height = NULL,
  boxToolSize = "sm", collapsible = TRUE, closable = FALSE)

Arguments

...
body content.

title
box title.

subtitle
box subtitle.

type
NULL by default. Choose the value 2 to try another skin.

background
Whether to enable a background image in the box header.

backgroundUrl
image url, if any. Background needs to be TRUE.

src
header image, if any (this is different of the background image).

color
background color: see here for a list of valid colors https://adminlte.io/themes/AdminLTE/pages/UI/general.html.

footer
box footer.

footer_padding
TRUE by default: whether the footer has margin or not.

width
box width (between 1 and 12).

height
box height.

boxToolSize
size of the toolbox: choose among "xs", "sm", "md", "lg".

collapsible
If TRUE, display a button in the upper right that allows the user to collapse the box.

closable
If TRUE, display a button in the upper right that allows the user to close the box.

Examples

if (interactive()) {
 library(shiny)
 library(shinydashboard)
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     widgetUserBox(
      title = "Nadia Carmichael",
      subtitle = "lead Developer",
      type = 2,
      src = "https://adminlte.io/themes/AdminLTE/dist/img/user7-128x128.jpg",
      color = "yellow",
      "Some text here!",
      footer = "The footer here!"
     ),
     widgetUserBox(
      title = "Alexander Pierce",
      subtitle = "Founder & CEO",
      type = NULL,
      src = "https://adminlte.io/themes/AdminLTE/dist/img/user1-128x128.jpg",
      color = "aqua-active",
      closable = TRUE,
      "Some text here!",
      footer = "The footer here!"
     ),
     widgetUserBox(
      title = "Elizabeth Pierce",
      subtitle = "Web Designer",
      type = NULL,
      src = "https://adminlte.io/themes/AdminLTE/dist/img/user3-128x128.jpg",
      background = TRUE,
      backgroundUrl = "https://www.planwallpaper.com/static/images/744081-background-wallpaper.jpg",
      closable = TRUE,
      "Some text here!",
      footer = "The footer here!"
     )
    ),
    title = "widgetUserBox"
  ),
  server = function(input, output) { }
 )
}