Create a box for the main body of a dashboard

Boxes can be used to hold content in the main body of a dashboard.

box(
  ...,
  inputId = NULL,
  title = NULL,
  footer = NULL,
  status = NULL,
  solidHeader = FALSE,
  background = NULL,
  icon = NULL,
  gradient = FALSE,
  boxToolSize = "sm",
  width = 6,
  height = NULL,
  collapsible = FALSE,
  collapsed = FALSE,
  closable = FALSE,
  label = NULL,
  dropdownMenu = NULL,
  sidebar = NULL,
  footerPadding = TRUE
)

Arguments

...
Contents of the box.

inputId
Box unique id. updateBox target.

title
Optional title.

footer
Optional footer text.

status
The status of the item This determines the item's background color. Valid statuses are listed in validStatuses.

solidHeader
Should the header be shown with a solid color background?

background
If NULL (the default), the background of the box will be white. Otherwise, a color string. Valid colors are listed in validColors.

icon
Header icon. Displayed before title. Expect icon.

gradient
Whether to allow gradient effect for the background color. Default to FALSE.

boxToolSize
Size of the toolbox: choose among "xs", "sm", "md", "lg".

width
The width of the box, using the Bootstrap grid system. This is used for row-based layouts. The overall width of a region is 12, so the default valueBox width of 4 occupies 1/3 of that width. For column-based layouts, use NULL for the width; the width is set by the column that contains the box.

height
The height of a box, in pixels or other CSS unit. By default the height scales automatically with the content.

collapsible
If TRUE, display a button in the upper right that allows the user to collapse the box.

collapsed
If TRUE, start collapsed. This must be used with collapsible=TRUE.

closable
If TRUE, display a button in the upper right that allows the user to close the box.

label
Slot for boxLabel.

dropdownMenu
List of items in the boxtool dropdown menu. Use boxDropdown.

sidebar
Slot for boxSidebar.

footerPadding
TRUE by default: whether the footer has margin or not.

Examples

## Only run this example in interactive R sessions
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 library(shinydashboardPlus)

 shinyApp(
   ui = dashboardPage(
     dashboardHeader(),
     dashboardSidebar(),
     dashboardBody(
      box(
        title = "Closable Box with dropdown", 
        closable = TRUE, 
        width = 12,
        status = "warning", 
        solidHeader = FALSE, 
        collapsible = TRUE,
        label = boxLabel(
         text = 1,
         status = "danger",
         style = "circle"
        ),
        dropdownMenu = boxDropdown(
         boxDropdownItem(url = "http://www.google.com", name = "Link to google"),
         boxDropdownItem(url = "#", name = "item 2"),
         dropdownDivider(),
         boxDropdownItem(url = "#", name = "item 3")
        ),
        sidebar = boxSidebar(
         startOpen = TRUE,
         inputId = "mycardsidebar",
         sliderInput(
          "obs", 
          "Number of observations:",
          min = 0, 
          max = 1000, 
          value = 500
         )
        ),
        plotOutput("distPlot")
       )
     )
   ),
   server = function(input, output) {
    output$distPlot <- renderPlot({
     hist(rnorm(input$obs))
    })
   }
 )
}