descriptionBlock {shinydashboardPlus}	R Documentation

AdminLTE2 description block

Description

Create a description block, perfect for writing statistics

Usage

descriptionBlock(number = NULL, number_color = NULL,
  number_icon = NULL, header = NULL, text = NULL,
  right_border = TRUE, margin_bottom = FALSE)

Arguments

number
any number.

number_color
number color: see here for a list of valid colors https://adminlte.io/themes/AdminLTE/pages/UI/general.html.

number_icon
number icon, if any. Should be written like "fa fa-times".

header
bold text.

text
additional text.

right_border
TRUE by default. Whether to display a right border to separate two blocks. The last block on the right should not have a right border.

margin_bottom
FALSE by default. Set it to TRUE when the descriptionBlock is used in a boxPad context.

Author(s)

David Granjon, dgranjon@ymail.com

Examples

if (interactive()) {
 library(shiny)
 library(shinydashboard)
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     box(
      solidHeader = FALSE,
      title = "Status summary",
      background = NULL,
      width = 4,
      status = "danger",
      footer = fluidRow(
        column(
          width = 6,
          descriptionBlock(
            number = "17%", 
            number_color = "green", 
            number_icon = "fa fa-caret-up",
            header = "$35,210.43", 
            text = "TOTAL REVENUE", 
            right_border = TRUE,
            margin_bottom = FALSE
          )
        ),
        column(
          width = 6,
          descriptionBlock(
            number = "18%", 
            number_color = "red", 
            number_icon = "fa fa-caret-down",
            header = "1200", 
            text = "GOAL COMPLETION", 
            right_border = FALSE,
            margin_bottom = FALSE
          )
        )
      )
     )
    ),
    title = "Description Blocks"
  ),
  server = function(input, output) { }
 )
}

[Package shinydashboardPlus version 0.7.0 Index]