dropdownBlock {shinydashboardPlus}	R Documentation

Create a dropdown block to place in a dashboard header

Description

Create a dropdown block to place in a dashboard header

Usage

dropdownBlock(..., id, icon = NULL, title = NULL,
  badgeStatus = "danger")

Arguments

...
Items to put in the menu.

id
Dropdown block id.

icon
An icon to display in the header. By default, the icon is automatically selected depending on type, but it can be overriden with this argument.

title
Dropdown block title.

badgeStatus
Dropdown badge status.

See Also

dashboardHeaderPlus for example usage.

[Package shinydashboardPlus version 0.7.0 Index]