gradientBox {shinydashboardPlus}	R Documentation

AdminLTE2 gradient box

Description

Create an enhanced box

Usage

gradientBox(..., title = NULL, icon = NULL, gradientColor = NULL,
  boxToolSize = "sm", footer = NULL, width = 6, height = NULL,
  collapsible = TRUE, closable = FALSE, footer_padding = TRUE)

Arguments

...
body content.

title
box title.

icon
header icon, if any. Should be written like "fa fa-times".

gradientColor
color of the box: see here for a list of valid colors https://adminlte.io/themes/AdminLTE/pages/UI/general.html.

boxToolSize
size of the toolbox: choose among "xs", "sm", "md", "lg".

footer
the box footer.

width
box width (between 1 and 12).

height
box height.

collapsible
If TRUE, display a button in the upper right that allows the user to collapse the box.

closable
If TRUE, display a button in the upper right that allows the user to close the box.

footer_padding
TRUE by default: whether the footer has margin or not.

Author(s)

David Granjon, dgranjon@ymail.com

Examples

if (interactive()) {
 library(shiny)
 library(shinydashboard)
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     gradientBox(
      title = "My gradient Box",
      icon = "fa fa-th",
      gradientColor = "teal", 
      boxToolSize = "sm", 
      footer = sliderInput(
       "obs", 
       "Number of observations:",
        min = 0, max = 1000, value = 500
       ),
      "This is a gradient box"
      ),
      gradientBox(
      title = "My gradient Box",
      icon = "fa fa-heart",
      gradientColor = "maroon", 
      boxToolSize = "xs", 
      closable = TRUE,
      footer = "The footer goes here. You can include anything",
      "This is a gradient box"
      )
    ),
    title = "gradientBox"
  ),
  server = function(input, output) { }
 )
}

[Package shinydashboardPlus version 0.7.0 Index]