renderUser {shinydashboardPlus}	R Documentation

Create dynamic user output (server side)

Description

Create dynamic user output (server side)

Usage

renderUser(expr, env = parent.frame(), quoted = FALSE, outputArgs = list())

Arguments

expr
An expression that returns a Shiny tag object, HTML(), or a list of such objects.

env
The environment in which to evaluate expr.

quoted
Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.

outputArgs
A list of arguments to be passed through to the implicit call to uiOutput() when renderUI is used in an interactive R Markdown document.

See Also

userOutput for the corresponding client side function and examples.

Other user outputs: userOutput()

[Package shinydashboardPlus version 2.0.0 Index]