renderUser {shinydashboardPlus}	R Documentation

Create dynamic user output (server side)

Description

Create dynamic user output (server side)

Usage

renderUser(expr, env = parent.frame(), quoted = FALSE, outputArgs = list())

Arguments

expr
An expression that returns a Shiny tag object, HTML(), or a list of such objects.

env
The parent environment for the reactive expression. By default, this is the calling environment, the same as when defining an ordinary non-reactive expression. If expr is a quosure and quoted is TRUE, then env is ignored.

quoted
If it is TRUE, then the quote()ed value of expr will be used when expr is evaluated. If expr is a quosure and you would like to use its expression as a value for expr, then you must set quoted to TRUE.

outputArgs
A list of arguments to be passed through to the implicit call to uiOutput() when renderUI is used in an interactive R Markdown document.

See Also

userOutput for the corresponding client side function and examples.

Other user outputs: userOutput()

[Package shinydashboardPlus version 2.0.4 Index]