timelineItem {shinydashboardPlus}	R Documentation

AdminLTE2 timeline item

Description

Create a timeline item

Usage

timelineItem(..., icon = NULL, color = NULL, time = NULL,
  title = NULL, border = TRUE, footer = NULL)

Arguments

...
any element such as timeLineItemMedia ...

icon
item icon such as "clock-o", "times", ...

color
item color: see here for a list of valid colors https://adminlte.io/themes/AdminLTE/pages/UI/general.html.

time
item date or time.

title
item title.

border
Whether to display a border between the header and the body. TRUE by default.

footer
item footer if any.

Author(s)

David Granjon, dgranjon@ymail.com

[Package shinydashboardPlus version 0.7.0 Index]