AdminLTE2 user box

Create user box

userBox(
  ...,
  title,
  subtitle = NULL,
  footer = NULL,
  color = NULL,
  width = 6,
  height = NULL,
  collapsible = TRUE,
  collapsed = FALSE,
  closable = FALSE,
  type = c(1, 2),
  image = NULL,
  backgroundImage = NULL,
  boxToolSize = "sm",
  footerPadding = TRUE
)

Arguments

...
body content.

title
box title.

subtitle
box subtitle.

footer
box footer.

color
background color: see here for a list of valid colors https://adminlte.io/themes/AdminLTE/pages/UI/general.html.

width
box width (between 1 and 12).

height
box height.

collapsible
If TRUE, display a button in the upper right that allows the user to collapse the box.

collapsed
If TRUE, start collapsed. This must be used with collapsible=TRUE.

closable
If TRUE, display a button in the upper right that allows the user to close the box.

type
User box type. Either 1 or 2. 1 corresponds to a centered user image, while 2 is a left aligned user image.

image
header image, if any (this is different of the background image).

backgroundImage
image url, if any. Background needs to be TRUE.

boxToolSize
size of the toolbox: choose among "xs", "sm", "md", "lg".

footerPadding
TRUE by default: whether the footer has margin or not.

Author

David Granjon, dgranjon@ymail.com

Examples

if (interactive()) {
 library(shiny)
 library(shinydashboard)
 library(shinydashboardPlus)
 
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     userBox(
      title = "Nadia Carmichael",
      subtitle = "lead Developer",
      type = 2,
      image = "https://adminlte.io/themes/AdminLTE/dist/img/user7-128x128.jpg",
      color = "yellow",
      "Some text here!",
      footer = "The footer here!"
     ),
     userBox(
      title = "Alexander Pierce",
      subtitle = "Founder & CEO",
      type = 1,
      image = "https://adminlte.io/themes/AdminLTE/dist/img/user1-128x128.jpg",
      color = "aqua",
      closable = TRUE,
      "Some text here!",
      footer = "The footer here!"
     ),
     userBox(
      title = "Elizabeth Pierce",
      subtitle = "Web Designer",
      image = "https://adminlte.io/themes/AdminLTE/dist/img/user3-128x128.jpg",
      backgroundImage = "https://avante.biz/wp-content/uploads/Minimal-Wallpaper/Minimal-Wallpaper-001.jpg",
      closable = TRUE,
      "Some text here!",
      footer = "The footer here!"
     )
    ),
    title = "UserBox"
  ),
  server = function(input, output) { }
 )
}