userPost {shinydashboardPlus}	R Documentation

AdminLTE2 user post

Description

Create a user post

Usage

userPost(..., id = NULL, src = NULL, author = NULL,
  description = NULL, collapsible = TRUE, collapsed = FALSE)

Arguments

...
post content, slot for userPostToolItemList.

id
unique id of the post.

src
profile image, if any.

author
post author.

description
post description.

collapsible
If TRUE, display a button in the upper right that allows the user to collapse the comment.

collapsed
Whether the comment is collapsed when the application starts, FALSE by default.

Author(s)

David Granjon, dgranjon@ymail.com

Examples

if (interactive()) {
 library(shiny)
 library(shinydashboard)
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     box(
      title = "Box with user comment",
      status = "primary",
      userPost(
       id = 1,
       src = "https://adminlte.io/themes/AdminLTE/dist/img/user1-128x128.jpg",
       author = "Jonathan Burke Jr.",
       description = "Shared publicly - 7:30 PM today",
       "Lorem ipsum represents a long-held tradition for designers, 
       typographers and the like. Some people hate it and argue for 
       its demise, but others ignore the hate as they create awesome 
       tools to help create filler text for everyone from bacon 
       lovers to Charlie Sheen fans.",
       userPostToolItemList(
        userPostToolItem(dashboardLabel("item 1")),
        userPostToolItem(dashboardLabel("item 2", status = "danger"), side = "right")
       )
      ),
      userPost(
       id = 2,
       src = "https://adminlte.io/themes/AdminLTE/dist/img/user6-128x128.jpg",
       author = "Adam Jones",
       description = "Shared publicly - 5 days ago",
       userPostMedia(src = "https://adminlte.io/themes/AdminLTE/dist/img/photo2.png"),
       userPostToolItemList(
        userPostToolItem(dashboardLabel("item 1")),
        userPostToolItem(dashboardLabel("item 2", status = "danger"), side = "right")
       )
      )
     )
    ),
    title = "userPost"
  ),
  server = function(input, output) { }
 )
}

[Package shinydashboardPlus version 0.7.0 Index]